+++ docs/pull_directory/CHANGELOG.md
@@ -2,6 +2,17 @@
 
 All notable changes to this gear are documented in this file.
 
+## 2.0.4
+
+* Add logging when directory records are ignored.
+* Change logging of ignored data types
+
+## 2.0.2
+
+* Update directory authorization model to match directory record.
+* Use email if auth email is not given.
+* Only export non-None fields.
+  
 ## 2.0.0
 
 * Adds this CHANGELOG